Header handling, removed handle name restriction
The code restructure could break header handling when using the pre-made
response methods (by sending a header before send_response). This is
fixed now by correctly adding headers in start_response.

The pre-modular request routing forbade requests with a _ in it for
safety reasons. This is no longer an issue, so that restriction has
been removed.